SoftCollage/requirements.txt
argparse
opencv-python
torch
numpy
pillow
matplotlib
torchvision
scipy